moved intensity measurement along vectors to `measurements` module
