+++ src/napari_stress/_measurements/__init__.py
@@ -23,3 +23,4 @@ from .geodesics import (geodesic_distance_matrix,
 from .stresses import anisotropic_stress, tissue_stress_tensor, maximal_tissue_anisotropy
 from .toolbox import comprehensive_analysis
 from .measurements import distance_to_k_nearest_neighbors
+from .intensity import sample_intensity_along_vector
